+++ src/main/java/com/jnape/palatable/lambda/functor/Cartesian.java
@@ -5,15 +5,16 @@ import com.jnape.palatable.lambda.adt.hlist.Tuple2;
 import java.util.function.Function;
 
 /**
- * {@link Profunctor} strength in the cartesian sense: <code>p a b -&gt; p (c, a) (c, b)</code> for any type
+ * {@link Profunctor} strength in the cartesian product sense: <code>p a b -&gt; p (c ^ a) (c ^ b)</code> for any type
  * <code>c</code>.
  *
  * @param <A> the type of the left parameter
  * @param <B> the type of the left parameter
- * @param <S> the unification parameter
+ * @param <P> the unification parameter
  * @see com.jnape.palatable.lambda.functions.Fn1
+ * @see Cocartesian
  */
-public interface Cartesian<A, B, S extends Cartesian<?, ?, S>> extends Profunctor<A, B, S> {
+public interface Cartesian<A, B, P extends Cartesian<?, ?, P>> extends Profunctor<A, B, P> {
 
     /**
      * Pair some type <code>C</code> to this profunctor's carrier types.
@@ -21,7 +22,7 @@ public interface Cartesian<A, B, S extends Cartesian<?, ?, S>> extends Profuncto
      * @param <C> the paired type
      * @return the cartesian-strengthened profunctor
      */
-    <C> Cartesian<Tuple2<C, A>, Tuple2<C, B>, S> cartesian();
+    <C> Cartesian<Tuple2<C, A>, Tuple2<C, B>, P> cartesian();
 
     /**
      * Pair the covariantly-positioned carrier type with the contravariantly-positioned carrier type. This can be
@@ -29,25 +30,25 @@ public interface Cartesian<A, B, S extends Cartesian<?, ?, S>> extends Profuncto
      *
      * @return the profunctor with the first parameter carried
      */
-    default Cartesian<A, Tuple2<A, B>, S> carry() {
+    default Cartesian<A, Tuple2<A, B>, P> carry() {
         return this.<A>cartesian().contraMap(Tuple2::fill);
     }
 
     @Override
-    <Z, C> Cartesian<Z, C, S> diMap(Function<? super Z, ? extends A> lFn, Function<? super B, ? extends C> rFn);
+    <Z, C> Cartesian<Z, C, P> diMap(Function<? super Z, ? extends A> lFn, Function<? super B, ? extends C> rFn);
 
     @Override
-    default <Z> Cartesian<Z, B, S> diMapL(Function<? super Z, ? extends A> fn) {
-        return (Cartesian<Z, B, S>) Profunctor.super.<Z>diMapL(fn);
+    default <Z> Cartesian<Z, B, P> diMapL(Function<? super Z, ? extends A> fn) {
+        return (Cartesian<Z, B, P>) Profunctor.super.<Z>diMapL(fn);
     }
 
     @Override
-    default <C> Cartesian<A, C, S> diMapR(Function<? super B, ? extends C> fn) {
-        return (Cartesian<A, C, S>) Profunctor.super.<C>diMapR(fn);
+    default <C> Cartesian<A, C, P> diMapR(Function<? super B, ? extends C> fn) {
+        return (Cartesian<A, C, P>) Profunctor.super.<C>diMapR(fn);
     }
 
     @Override
-    default <Z> Cartesian<Z, B, S> contraMap(Function<? super Z, ? extends A> fn) {
-        return (Cartesian<Z, B, S>) Profunctor.super.<Z>contraMap(fn);
+    default <Z> Cartesian<Z, B, P> contraMap(Function<? super Z, ? extends A> fn) {
+        return (Cartesian<Z, B, P>) Profunctor.super.<Z>contraMap(fn);
     }
 }